poly16z/pyproject.toml
[project]
name = "poly16z"
version = "0.1.0"
description = "AI-Powered Polymarket Trading Framework"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.3.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "py-clob-client>=0.18.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
